Fix endpoints on dashboard

+++ pages/api/v1/articles/delete/index.ts
@@ -1,8 +1,9 @@
 import { NextApiRequest, NextApiResponse } from "next";
 
 import { db } from "../../../../../mongo/client";
-import { validateJwt } from "../../../../../src/helpers/auth/jwt";
+import { validateClientJwt } from "../../../../../src/helpers/auth/jwt";
 import Article from "../../../../../src/models/Article";
+import { DataResponse } from "../../../../../types/types";
 
 type JsonResponse = {
   ok: boolean;
@@ -10,35 +11,56 @@ type JsonResponse = {
   data: unknown;
 };
 
-export default validateJwt(async function updateArticles(
+export default async function deleteAllArticles(
   req: NextApiRequest,
   res: NextApiResponse<JsonResponse>,
 ) {
-  if (req.method !== "DELETE") {
-    return res.status(405).json({
-      message: "Method not allowed",
-      ok: false,
-      data: null,
-    });
-  }
-
-  if (req.method === "DELETE") {
-    db(process.env.MONGO_URI, res);
-
-    const allArticles = await Article.deleteMany({});
-
-    if (!allArticles) {
-      return res.status(404).json({
-        message: "For some reason, articles were not deleted",
+  return new Promise(async (resolve) => {
+    if (req.method !== "DELETE") {
+      return res.status(405).json({
+        message: "Method not allowed",
         ok: false,
         data: null,
       });
     }
 
-    return res.status(200).json({
-      message: "Articles deleteded",
-      ok: true,
-      data: null,
-    });
-  }
-});
+    if (req.method === "DELETE") {
+      const token = req?.headers?.authorization?.split(" ")[1] as string;
+
+      const checkToken = (await validateClientJwt(token)) as DataResponse;
+
+      if (!checkToken.ok) {
+        return resolve(
+          res.status(401).json({
+            message: "Invalid token",
+            ok: false,
+            data: null,
+          }),
+        );
+      }
+
+      db(process.env.MONGO_URI, res);
+      Article.deleteMany({})
+        .then((data) => {
+          return resolve(
+            res.status(200).json({
+              message: "Article deleteded",
+              ok: true,
+              data: data,
+            }),
+          );
+        })
+        .catch((error) => {
+          console.log(error);
+
+          return resolve(
+            res.status(500).json({
+              message: "Internal server error",
+              ok: false,
+              data: null,
+            }),
+          );
+        });
+    }
+  });
+}
